fix(*): split functionality into multiple modules, add typedoc
