+++ package.json
@@ -11,7 +11,8 @@
     "cypress": "./node_modules/.bin/cypress open",
     "prepublishOnly": "cp -r ./build/* . && rm -rf ./build",
     "postpublish": "git clean -fd",
-    "semantic-release": "semantic-release"
+    "semantic-release": "semantic-release",
+    "typedoc": "typedoc"
   },
   "release": {
     "plugins": [
@@ -46,6 +47,7 @@
     "tslint": "^6.1.2",
     "tslint-config-prettier": "^1.18.0",
     "typedoc": "^0.17.7",
+    "typedoc-neo-theme": "^1.0.8",
     "typescript": "^3.9.5",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.43.0",
