+++ src/types/SASjsConfig.ts
@@ -1,0 +1,8 @@
+export class SASjsConfig {
+  serverUrl: string = "";
+  pathSAS9: string = "";
+  pathSASViya: string = "";
+  appLoc: string = "";
+  serverType: string = "";
+  debug: boolean = true;
+}
