+++ src/types/SASjsRequest.ts
@@ -1,0 +1,8 @@
+export interface SASjsRequest {
+  serviceLink: string;
+  timestamp: Date;
+  sourceCode: string;
+  generatedCode: string;
+  logFile: string;
+  SASWORK: any;
+}
