+++ src/types/SASjsWaitingRequest.ts
@@ -1,0 +1,10 @@
+export interface SASjsWatingRequest {
+  requestPromise: {
+    promise: any;
+    resolve: any;
+    reject: any;
+  };
+  programName: string;
+  data: any;
+  params?: any;
+}
