+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./SASjsConfig";
+export * from "./SASjsRequest";
+export * from "./SASjsWaitingRequest";
