+++ src/utils/compareTimestamps.ts
@@ -1,0 +1,5 @@
+import { SASjsRequest } from "../types/SASjsRequest";
+
+export const compareTimestamps = (a: SASjsRequest, b: SASjsRequest) => {
+  return b.timestamp.getTime() - a.timestamp.getTime();
+};
