+++ src/utils/isAuthorizeFormRequired.ts
@@ -1,0 +1,3 @@
+export const isAuthorizeFormRequired = (response: string) => {
+  return /<form.+action="(.*Logon\/oauth\/authorize[^"]*).*>/gm.test(response);
+};
