+++ src/utils/isLoginRequired.ts
@@ -1,0 +1,5 @@
+export const isLogInRequired = (response: string): boolean => {
+  const pattern: RegExp = /<form.+action="(.*Logon[^"]*).*>/gm;
+  const matches = pattern.test(response);
+  return matches;
+};
