+++ src/utils/isLoginSuccess.ts
@@ -1,0 +1,2 @@
+export const isLogInSuccess = (response: string): boolean =>
+  /You have signed in/gm.test(response);
