+++ src/utils/serialize.ts
@@ -1,0 +1,15 @@
+export const serialize = (obj: any) => {
+  const str: any[] = [];
+  for (const p in obj) {
+    if (obj.hasOwnProperty(p)) {
+      if (obj[p] instanceof Array) {
+        for (let i = 0, n = obj[p].length; i < n; i++) {
+          str.push(encodeURIComponent(p) + "=" + encodeURIComponent(obj[p][i]));
+        }
+      } else {
+        str.push(encodeURIComponent(p) + "=" + encodeURIComponent(obj[p]));
+      }
+    }
+  }
+  return str.join("&");
+};
